Move database query functions into class
